The Atmos Mega-Update
https://github.com/space-wizards/space-station-14/pull/36124
https://github.com/space-wizards/space-station-14/pull/36910
https://github.com/space-wizards/space-station-14/pull/38244
https://github.com/space-wizards/space-station-14/pull/37881
https://github.com/space-wizards/space-station-14/pull/37385
https://github.com/space-wizards/space-station-14/pull/38672
https://github.com/space-wizards/space-station-14/pull/36708
https://github.com/space-wizards/space-station-14/pull/38974
https://github.com/space-wizards/space-station-14/pull/37341
https://github.com/space-wizards/space-station-14/pull/33837
https://github.com/space-wizards/space-station-14/pull/33800
https://github.com/space-wizards/space-station-14/pull/38075
https://github.com/space-wizards/space-station-14/pull/37531
https://github.com/space-wizards/space-station-14/pull/37568
https://github.com/space-wizards/space-station-14/pull/38010

https://github.com/space-wizards/RobustToolbox/pull/5932
https://github.com/space-wizards/RobustToolbox/pull/5602
https://github.com/space-wizards/RobustToolbox/pull/5841

I just wanted pipe layering.  I wasn't expecting it to become an odyssey.  But here we are, many hours later, with an absolutely MASSIVE update that brings us close to upstream in terms of Atmospherics, with lots of back-end changes to smooth out the experience.  There are new toys of course, like the input pressure regulator and of course pipe layering and gas manifolds.  There's also the spraypainter.  It's amazing now, go make one and paint everything.  Oh, and air grenades, you can buy them in Cargo.

There's probably more things I forgot and because the old method of spriting is now considered obsolete compiling things generates a few hundred more warnings, but it seems to be functional, somehow.