Updated Gearharness Image and added Transparent Shoes
Updated the Coyote Bayou Nudist Gear Harness with an image of a gear harness instead of a paw. Lowered the loadout price to the current vendor price of 100. Added Transparent foot wraps for those of us who like to see our paws, but still allowing us to wear shoes. Gave them the same price as regular Foot wraps.